cleaned up a bit

+++ src/goboy/opcodes.go
@@ -823,7 +823,7 @@ func (gb *GameBoy) ADD_sp_e(ins [2]uint8) {
 	aVal := gb.get16Reg(SP)
 	bVal := uint16(int8(ins[1]))
 	out := aVal + bVal
-	gb.set16Reg(SP, uint16(out))
+	gb.set16Reg(SP, out)
 	gb.modifyFlag(N_FLAG, CLEAR)
 	gb.modifyFlag(Z_FLAG, CLEAR)
 	if (aVal&0x0fff)+(bVal&0x0fff) > 0x0fff {
